RC2014: change master/slave rule
When the 2nd NEXTOR kernel is loaded, the rules for identifing
and selecting master/slave kernels has changed slightly.

Whereas before, if the 2 kernels had the same main and major
values then the kernel identified as master was left unchanged.

Now, if they are the same, the 2nd loaded kernel will promote
itself to master.

example:

If in slot 2 - we have kernel 2.1.0
and in slot 3, we have kernel 2.1.1

Previously slot 2 would become master - as only the main and
major version numbers are compared.  The 2nd kernel in slot 3
would only promote itself to master, if its main and major
numbers are greater.

Now, the comparison will cause the kernel in slot 3 to
promote to master, if both main and major version are the same
or greater.